+++ R/smovie.R
@@ -35,6 +35,8 @@
 #' \itemize{
 #'   \item {\link[=clt]{Central Limit Theorem: sampling distribution
 #'     of a sample mean}}
+#'   \item {\link[=mean_vs_median]{Sampling distributions of the mean
+#'     and median}}
 #'   \item {\link[=ett]{Extremal Types Theorem: sampling distribution
 #'     of a sample maximum}}
 #'   \item {\link[=correlation]{Pearson product moment correlation